% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchScale.R
\name{BatchScale}
\alias{BatchScale}
\title{Scale taxonomy summary plots (output of summarize_taxa_through_plots.py of QIIME) by biomass}
\usage{
BatchScale(biomass_file, taxonomy_directory, OutputPrefix = "Scaled",
  XAxisLabel = "Timepoint", YAxisLabel = "ug DNA per mg Fecal Pellet",
  OutputDirectory = ".", batchName = "Sample", ...)
}
\arguments{
\item{biomass_file}{input file that contains biomass data for each x-axis category of your summary plots}

\item{taxonomy_directory}{path to directory that contains output of the QIIME funciton call}

\item{OutputPrefix}{OPTIONAL: a string to append to the beginning of output files (default = 'Scaled')}

\item{XAxisLabel}{OPTIONAL: a string to label x-axis of plots (default = 'Timepoint')}

\item{YAxisLabel}{OPTIONAL: a string to label y-axis of plots (default = 'ug DNA per mg Fecal Pellet')}

\item{OutputDirectory}{OPTIONAL: path to output directory (by default './')}

\item{batchName}{OPTIONAL: string to label batch of plots (default = 'Sample')}

\item{...}{Optional arguments}
}
\value{
a list that contains within it: a list of the scale taxonomy plots (at the various levels), and the scaled plots
}
\description{
Scale taxonomy summary plots (output of summarize_taxa_through_plots.py of QIIME) by biomass
}

