% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassutils.R
\docType{package}
\name{biomassutils}
\alias{biomassutils}
\alias{biomassutils-package}
\title{biomassutils: A Helper Package for Microbial Biomass Data Analysis}
\description{
This package helps generate and analyze microbiome biomass data, prepare sequencing of microbiome samples, and analye sequencing data.
}
\section{Data Generation & Analysis}{

Functions here
}

\section{Sequencing Prep}{

Functions here
}

\section{Sequencing Data Analysis}{

Functions here
}

