% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomassWorkflow.R
\docType{package}
\name{BiomassWorkflow}
\alias{BiomassWorkflow}
\alias{BiomassWorkflow-package}
\title{BiomassWorkflow: A Helper Package for Microbial Biomass Data Analysis}
\description{
This package helps generate and analyze microbiome biomass data, prepare sequencing of microbiome samples, and analye sequencing data.
}

