% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_biom_friendly.R
\name{convert_to_biom_friendly}
\alias{convert_to_biom_friendly}
\title{Function to convert scaled/unscaled output of individual_scale to biom-friendly format}
\usage{
convert_to_biom_friendly(table, output_name, scale_factor = 1e+06)
}
\arguments{
\item{table}{table from scaled/unscaled output of individual_scale}

\item{output_name}{name for output file (written)}

\item{scale_factor}{default = 1e6 -value to multiply all smaples by to generate "reads" that are scaled appropriately)}
}
\value{
output table
}
\description{
Function to convert scaled/unscaled output of individual_scale to biom-friendly format
}

