% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_lefse.R
\name{convert_to_lefse}
\alias{convert_to_lefse}
\title{Convert to lefse-friendly}
\usage{
convert_to_lefse(table, sample = "X.SampleID", class = "Condition",
  subclass = NULL, depth = "Species")
}
\arguments{
\item{table}{table from individual_scale output (compact_by_taxonomy), at a given taxonomy}

\item{sample}{string to define column that contains sample ID's}

\item{class}{string to define column that contains class groupings}

\item{subclass}{string to define column that contains class groupings (optional)}

\item{depth}{string defining taxonomic depth to go to (optional - default = 'Species')}
}
\value{
table that you can save as tab delimited file and pipe to LEFSE
}
\description{
Convert to lefse-friendly
}

