% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_phyloseq.R
\name{convert_to_phyloseq}
\alias{convert_to_phyloseq}
\title{Function to convert sequencing object to phyloseq object with optional addition of phylogenetic tree and taxonomy data}
\usage{
convert_to_phyloseq(sequencing_object, treefile = NULL, taxonomy = NULL)
}
\arguments{
\item{sequencing_object}{output of individual_scale function (will contian sample metadata)}

\item{treefile}{(optional) greengenes taxonomic tree file}

\item{taxonomy}{(optional) greengenes taxonomic information file}
}
\value{
phyloseq object with provided data
}
\description{
Function to convert sequencing object to phyloseq object with optional addition of phylogenetic tree and taxonomy data
}
