% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD_DataAnalysis.R
\name{Deprecated_DataAnalysis}
\alias{Deprecated_DataAnalysis}
\title{Deprecated function for Data Analysis. See DataAnalysis instead}
\usage{
Deprecated_DataAnalysis(plate_reader_csv_file, mapping_csv_file, exp_id,
  standards_plate_reader_csv_file = plate_reader_csv_file,
  standards_mapping_csv_file = mapping_csv_file, ...)
}
\arguments{
\item{plate_reader_csv_file}{plate reader file (old version)}

\item{mapping_csv_file}{mapping file}

\item{exp_id}{experiment id}

\item{standards_plate_reader_csv_file}{plate reader file for plate containing standards (from old machine)}

\item{standards_mapping_csv_file}{mapping file for plate contianing standards (for wells with standards, Type must be 'Standard')}

\item{...}{other arguments}
}
\value{
data table with biomass data calculated
}
\description{
Deprecated function for Data Analysis. See DataAnalysis instead
}

