% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD_ParsePlateReaderFile.R
\name{Deprecated_ParsePlateReaderFile}
\alias{Deprecated_ParsePlateReaderFile}
\title{Deprecated plate reader file parser}
\usage{
Deprecated_ParsePlateReaderFile(plate_reader_file)
}
\arguments{
\item{plate_reader_file}{plate reader file from old plate reader}
}
\value{
list that includes data in table format and the number of measurements that the plate reader took per sample
}
\description{
Deprecated plate reader file parser
}

