% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD_StandardAnalysis.R
\name{Deprecated_StandardAnalysis}
\alias{Deprecated_StandardAnalysis}
\title{Deprecated function to process standards information and create line of best fit}
\usage{
Deprecated_StandardAnalysis(standards_plate_reader_csv_file,
  standards_mapping_csv_file, exp_id, print = FALSE, ...)
}
\arguments{
\item{standards_plate_reader_csv_file}{plate reader file for plate containing standards (from old machine)}

\item{standards_mapping_csv_file}{mapping file for plate contianing standards (for wells with standards, Type must be 'Standard')}

\item{exp_id}{experiment id}

\item{print}{print standards information or not}

\item{...}{...}
}
\value{
list containing a table of the standards, and the information for the standard curve
}
\description{
Deprecated function to process standards information and create line of best fit
}
