% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropcolumns.R
\name{dropcolumns}
\alias{dropcolumns}
\title{Simple function to remove columns that are only comprised of NA's}
\usage{
dropcolumns(table)
}
\arguments{
\item{table}{table to remove NAs from}
}
\value{
output table with no columns that are only NAs (trims columns)
}
\description{
Simple function to remove columns that are only comprised of NA's
}

