% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_legend}
\alias{get_legend}
\title{Create a stand-alone legend for plotting}
\usage{
get_legend(ggplot_item)
}
\arguments{
\item{ggplot_item}{ggplot item}
}
\value{
legend
}
\description{
Create a stand-alone legend for plotting
}

