% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fecal_IgA.R
\name{IgAAnalysis}
\alias{IgAAnalysis}
\title{Function to process Fecal IgA Elisa data}
\usage{
IgAAnalysis(plate_reader_file, mapping_file, shiny = FALSE, type = NULL,
  print = FALSE, ...)
}
\arguments{
\item{plate_reader_file}{the file path of the plate reader file (in .csv or .xls(x) format)}

\item{mapping_file}{mapping file for plate (for wells with standards, Type must be 'Standard')}

\item{shiny}{OPTIONAL: necessary for running with shiny app interface since filenames are not the same.}

\item{type}{OPTIONAL: necessary for running with shiny app, must specify file type}

\item{print}{OPTIONAL: whether or not to save a copy of the standards curve to the working directory}

\item{...}{optional inputs}
}
\value{
list containing a table of the standards, and the information for the standard curve
}
\description{
Function to process Fecal IgA Elisa data
}

