% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_biom_tables.R
\name{make_biom_tables}
\alias{make_biom_tables}
\title{Make biom tables (relative abundance and absolute abundance) from a sequencing object (individual_scale output)}
\usage{
make_biom_tables(sequencing_object, output_type = "biom",
  output_prefix = NULL, output_directory = getwd())
}
\arguments{
\item{sequencing_object}{output from individual_scale that contains microbial density-scaled sequencing data}

\item{output_type}{'biom' or 'phyloseq', to indicate which format to generate the resulant OTU tables in within R}

\item{output_prefix}{(optional) prefix to give otu tables. By providing this, you will write .biom otu tables to your working directory or specified output directory}

\item{output_directory}{(optional) specify an output directory to write otu tables in .biom format to (must provide output_prefix)}
}
\value{
a list of otu tables (relative and absolute) in the specified format. Additionally, if specified, .biom files will be written
}
\description{
Make biom tables (relative abundance and absolute abundance) from a sequencing object (individual_scale output)
}

