% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map1}
\alias{map1}
\title{Mapping file for the first of two plates in a sample experiment}
\format{A table with 96 rows and 10 columns}
\source{
Generated in lab for the purposes of analyzing this type of data
}
\usage{
map1
}
\description{
A dataset that contains the information from a mapping, and is used to re-create a template mapping file. This data is used in the sample_file_maker script to generate a sample mapping file
}
\keyword{datasets}
