% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate_sequencing.R
\name{ordinate_sequencing}
\alias{ordinate_sequencing}
\title{Ordination of scaled sequencing object}
\usage{
ordinate_sequencing(sequencing_object, method = "euclidean")
}
\arguments{
\item{sequencing_object}{output of individual_scale function}

\item{method}{method for ordination}
}
\value{
a list of outputs containing ordination results and first 10 PC's
}
\description{
Ordination of scaled sequencing object
}

