% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Other_Data}
\alias{Other_Data}
\title{Fully processed other data (.csv)}
\format{Table with many samples}
\source{
Generated in lab for the purposes of analyzing this type of data
}
\usage{
Other_Data
}
\description{
A dataset that contains processed data from othere experiments
}
\keyword{datasets}
