% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_themes.R
\name{paper_theme}
\alias{paper_theme}
\title{Theme function for plotting}
\usage{
paper_theme(base_size = 8, base_family = "sans", bgcolor = "default")
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{base text family}

\item{bgcolor}{background color}
}
\description{
Theme function for plotting
}
