% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_themes.R
\name{paper_theme_tilted}
\alias{paper_theme_tilted}
\title{Theme function for plotting with tilted axes}
\usage{
paper_theme_tilted(base_size = 8, base_family = "sans",
  bgcolor = "default")
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{base text family}

\item{bgcolor}{background color}
}
\description{
Theme function for plotting with tilted axes
}

