% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMappingFile.R
\name{ParseMappingFile}
\alias{ParseMappingFile}
\title{Function to read in mapping file and tweak format so that it is friendly with the rest of the package functinos}
\usage{
ParseMappingFile(mapping_file)
}
\arguments{
\item{mapping_file}{string containing path to mapping file}
}
\value{
Mapping information in a data frame
}
\description{
Function to read in mapping file and tweak format so that it is friendly with the rest of the package functinos
}
