% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{quick conversion to percent}
\usage{
percent(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{number to convrt}

\item{digits}{number of digits}

\item{format}{format}

\item{...}{...}
}
\value{
percent formatted
}
\description{
quick conversion to percent
}
