% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJC_colors.R
\docType{data}
\name{phyla_colors}
\alias{phyla_colors}
\title{Color Scheme function for phyla}
\format{An object of class \code{character} of length 90.}
\usage{
phyla_colors
}
\description{
Color Scheme function for phyla
}
\keyword{datasets}

