% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_to_table.R
\name{plate_to_table}
\alias{plate_to_table}
\title{Plate to table}
\usage{
plate_to_table(plate_layout, size = 96)
}
\arguments{
\item{plate_layout}{file with plate layout (with headers)}

\item{size}{96 wells or 384 (defaults to 96)}
}
\value{
samples in table format
}
\description{
Plate to table
}
