% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateParser.R
\name{PlateParser}
\alias{PlateParser}
\title{Function to read in mapping file and tweak format so that it is friendly with the rest of the package functinos}
\usage{
PlateParser(plate_reader_file, num_reads = 1)
}
\arguments{
\item{plate_reader_file}{Fluorescence data from the plate reader (can be in .csv or .xls(x) format -- if in excel file, only the FIRST sheet within the file will be read)}

\item{num_reads}{OPTIONAL: number of measurements that the plate reader took for each sample (default = 1)}
}
\value{
a list containing: 1) a data frame with the data from the plate reader file and 2) the number of measurements taken for each sample
}
\description{
Function to read in mapping file and tweak format so that it is friendly with the rest of the package functinos
}

