% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotter.R
\name{plotter}
\alias{plotter}
\title{Function to generate an interactive data plotter}
\usage{
plotter(data)
}
\arguments{
\item{data}{a data frame to be plotted}
}
\value{
launches a shiny app that can help quickly visualize data. Has an interactive x and y components, and can toggle between scatter and box plots
}
\description{
Function to generate an interactive data plotter
}
