% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_label_pairings.R
\name{read_label_pairings}
\alias{read_label_pairings}
\title{Read Label Pairings}
\usage{
read_label_pairings(label_barcode_pairings, col1_name = "col1",
  col2_name = "col2", save = F)
}
\arguments{
\item{label_barcode_pairings}{File that contains scanned barcodes and labels}

\item{col1_name}{optional name of first item scanned (i.e. tube barcode or label barcode)}

\item{col2_name}{optional name of second item scanned (i.e. label barcode or tube barcode)}

\item{save}{optional to save resulting table to file in current working directory}
}
\value{
table with pairings
}
\description{
Read Label Pairings
}

