% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_file_maker.R
\name{sample_file_maker}
\alias{sample_file_maker}
\title{Function to generate sample data files to work wtih locally}
\usage{
sample_file_maker(directory = getwd())
}
\arguments{
\item{directory}{If given, a directory into which to create files}
}
\value{
A collection of files useful for getting started with the BiomassWorkflow package
}
\description{
Function to generate sample data files to work wtih locally
}
