% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_and_visualize.R
\name{scale_and_visualize}
\alias{scale_and_visualize}
\title{Function to perform individual_scale function AND vizualization functon}
\usage{
scale_and_visualize(biom_file, metadata_file, taxonomy_file, filter = 0)
}
\arguments{
\item{biom_file}{.biom table in rhdf5 format. It is wise to use a pre-filtered table (for example, de-duplicated)}

\item{metadata_file}{the mapping file used in QIIME to split your sequencing library (must include biomass data)}

\item{taxonomy_file}{a file containing taxonomy (e.g. output from assign_taxonomy.py in QIIME, or greengenes taxonomy)}

\item{filter}{OPTIONAL: minimum relative abundance to filter from data}
}
\value{
a list containing OTU tables with and without taxonomy (at each given depth), both scaled and unscaled (the output is a list of lists).
}
\description{
Function to perform individual_scale function AND vizualization functon
}

