% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_analysis_wrapper.R
\name{sequencing_analysis_wrapper}
\alias{sequencing_analysis_wrapper}
\title{Sequencing Analysis Wrapper}
\usage{
sequencing_analysis_wrapper(biom_file, metadata_file, taxonomy_file, tree_file,
  method = "PCoA", distance = "unifrac")
}
\arguments{
\item{biom_file}{otu table in biom format}

\item{metadata_file}{mapping file (QIIME-ready)}

\item{taxonomy_file}{file that contains taxonomy (e.g. greengenes)}

\item{tree_file}{taxonomic tree information (e.g. greengenes)}

\item{method}{dimensionality reduction method (e.g. PCoA, NMDS, DCA)}

\item{distance}{distance metric to use (e.g. unifrac, bray, jaccard)}
}
\value{
list containing individual components of analyzed data ready for plotting, etc.
}
\description{
Sequencing Analysis Wrapper
}

