% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_data_plotter.R
\name{sequencing_data_plotter}
\alias{sequencing_data_plotter}
\title{Make abundance bar plots of sequencing data}
\usage{
sequencing_data_plotter(sequencing_object, depth = "Phylum",
  x.var = "X.SampleID", abundance = "relative", wrap.groups = NULL,
  x.groups = NULL, y.groups = NULL, facet.scales = "free",
  facet.space = "free", colors = NULL, tilt.axis = T, x.axis.factor = F)
}
\arguments{
\item{sequencing_object}{sequencing object (e.g. result of individual_scale)}

\item{depth}{taxonomic depth to plot}

\item{x.var}{variable for x axis}

\item{abundance}{relative or absolute}

\item{x.groups}{grouping variable for faceting along x axis}

\item{y.groups}{grouping variable for faceting along y axis}

\item{facet.scales}{scales for facets (e.g. free)}

\item{facet.space}{space for facets (e.g. free)}

\item{colors}{colors to use for bars}

\item{tilt.axis}{tilt x axis?}

\item{x.axis.factor}{Make x axis variable factor?}
}
\value{
ggplot object
}
\description{
Make abundance bar plots of sequencing data
}
