% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_data_viz.R
\name{sequencing_data_viz}
\alias{sequencing_data_viz}
\title{Shiny App that aids in vizualizing sequencing objects returned from individual_scale}
\usage{
sequencing_data_viz(sequencing_object)
}
\arguments{
\item{sequencing_object}{output object of individual_scale}
}
\value{
launches shiny app that helps interactively visualize data
}
\description{
Shiny App that aids in vizualizing sequencing objects returned from individual_scale
}

