% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_prep_metagenomics.R
\name{sequencing_prep_metagenomics}
\alias{sequencing_prep_metagenomics}
\title{Function to produce files necessary to generate metagenomic sequencing runs that utilizes the Beckmann robot for dilutions and creates sequencing mapping files in QIIME-ready format, from multiple experiment data sets.}
\usage{
sequencing_prep_metagenomics(experiment_data_list, n_barcode_plates,
  barcode_map_file = NULL, output_directory = getwd(), filter = T)
}
\arguments{
\item{experiment_data_list}{an input list that contains the data from the samples you want to sequence (ideally, the output from DataAnalysis, or combinations of them)}

\item{n_barcode_plates}{the number of barcode plates available (i.e. if you have 96 unique sequencing barcodes, this would be 1)}

\item{barcode_map_file}{a string of the path to a file containing the barcode sequences, linker primer sequences, and barcode plate and well locations (see example data)}

\item{output_directory}{OPTIONAL: a path to the directory for your desired output (Default = working directory)}

\item{filter}{OPTIONAL: a boolean that indicates whether or not to filter out samples that have not passed the QC check from the DataAnalysis function (Default = TRUE)}
}
\value{
A list containing the data for as many sequencing runs necessary to sequence all of your samples. For each sequencing run, there will be a list of plates, corresponding to the data frames that contain the information necessary to use the Beckmann robot to dilute samples. There will also be a mapping file, and a list of samples included in that sequencing run.
}
\description{
Function to produce files necessary to generate metagenomic sequencing runs that utilizes the Beckmann robot for dilutions and creates sequencing mapping files in QIIME-ready format, from multiple experiment data sets.
}

