% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fecal_IgA.R
\name{sigmoid_model}
\alias{sigmoid_model}
\title{Helper function for IgA Elisa standard curve fitting}
\usage{
sigmoid_model(params, x)
}
\arguments{
\item{params}{parameters for model from nls fit}

\item{x}{data to be fit}
}
\value{
sigmoid model for standards
}
\description{
Helper function for IgA Elisa standard curve fitting
}
