% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_parser.R
\name{taxonomy_parser}
\alias{taxonomy_parser}
\title{Function to parse greengenes taxonomy file for use with phyloseq}
\usage{
taxonomy_parser(taxonomy_file)
}
\arguments{
\item{taxonomy_file}{string with path to taxonomy file (e.g. greengenes)}
}
\value{
taxonomy file suitable for use with phyloseq
}
\description{
Function to parse greengenes taxonomy file for use with phyloseq
}
