% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wet_Dry_Weight.R
\name{wet_dry_weight}
\alias{wet_dry_weight}
\title{Wet vs Dry Weight calculation}
\usage{
wet_dry_weight(empty_weights, full_weights, cooked_weights, mapping = NULL)
}
\arguments{
\item{empty_weights}{file with empty weights of tubes}

\item{full_weights}{file with weights of tubes with sample in them}

\item{cooked_weights}{file with weights of tubes after cooking at 105C overnight}

\item{mapping}{file to add metadata to samples}
}
\value{
data frame with wet and dry weight information
}
\description{
Wet vs Dry Weight calculation
}
