% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAnalysis.R
\name{DataAnalysis}
\alias{DataAnalysis}
\title{Main function call to generate microbial density data from DNA extraction protocol}
\usage{
DataAnalysis(plate_reader_file, mapping_csv_file, exp_id,
  standards_plate_reader_file = plate_reader_file,
  standards_mapping_csv_file = mapping_csv_file, volume = 2, scale = 7,
  shiny = FALSE, type_plate = NULL, type_standards_plate = NULL,
  print = FALSE, ...)
}
\arguments{
\item{plate_reader_file}{raw file from plate-reader (can be in .xls format, or .csv), but will only read in first sheet of excel file}

\item{mapping_csv_file}{a .csv file that contains information as to the location of the samples, the sample types, and their mass. A sample file has been included in this package and can be found at https://github.com/econtijoch/Biomass-Workflow/blob/master/SampleMapping.csv}

\item{exp_id}{a string that indicates the name of the experiment (used to label the plot of your standard curve)}

\item{standards_plate_reader_file}{OPTIONAL: if your standards are not on the same plate as your samples, you will need to provide the file path of the plate reader file for the standards (Default is to use same plate as samples)}

\item{standards_mapping_csv_file}{OPTIONAL: if your standards are not on the same plate as your samples, you will also need to provide the file path of a mapping file for the standards plate (Default is to use same plate as samples)}

\item{volume}{OPTIONAL: The volume (in uL) used to quantify DNA with Qubit (Default = 2)}

\item{scale}{OPTIONAL: The scale factor used in the DNA isolation protocol (Defualt = 7, which corresponds to taking 100uL of supernatant from the 700uL of the DIB added). For phenol extraction, use scale = 1.}

\item{shiny}{OPTIONAL: necessary for running with shiny app interface since filenames are not the same.}

\item{type_plate}{OPTIONAL: necessary for running with shiny app, must specify file type of plate}

\item{type_standards_plate}{OPTIONAL: necessary for running with shiny app, must specify type of plate for standards}

\item{print}{OPTIONAL: whether or not to save a copy of the standards curve to the working directory}

\item{...}{Optional arguments to pass}
}
\value{
output table with DNA concentrations, microbial density, metadata, and useful computations for downstream (sequencing) applications, and a plot of the standard curve
}
\description{
Main function call to generate microbial density data from DNA extraction protocol
}
