% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardAnalysis.R
\name{StandardAnalysis}
\alias{StandardAnalysis}
\title{Function to process standards information and create line of best fit}
\usage{
StandardAnalysis(standards_plate_reader_file, standards_mapping_csv_file,
  exp_id, shiny = FALSE, type = NULL, print = FALSE, ...)
}
\arguments{
\item{standards_plate_reader_file}{the file path of the plate reader file for the standards (in .csv or .xls(x) format)}

\item{standards_mapping_csv_file}{mapping file for plate contianing standards (for wells with standards, Type must be 'Standard')}

\item{exp_id}{experiment id}

\item{shiny}{OPTIONAL: necessary for running with shiny app interface since filenames are not the same.}

\item{type}{OPTIONAL: necessary for running with shiny app, must specify file type}

\item{print}{OPTIONAL: whether or not to save a copy of the standards curve to the working directory}

\item{...}{optional inputs}
}
\value{
list containing a table of the standards, and the information for the standard curve
}
\description{
Function to process standards information and create line of best fit
}

